requirements.txt
requests
future
python-dotenv